refactor: enum class 위치 변경
enum class는 constant라는 package를 만들어 관리했는데, 여러 class에서 enum class를 만들어 사용하기 때문에 각각 app과 global 바로 하위 package로 관리하는 것이 좋다고 생각이 들어 변경하였음.
