+++ src/main/java/com/zelusik/eatery/app/constant/member/Gender.java
@@ -1,4 +1,4 @@
-package com.zelusik.eatery.app.domain.constant;
+package com.zelusik.eatery.app.constant.member;
 
 public enum Gender {
