+++ src/main/java/com/zelusik/eatery/app/constant/member/LoginType.java
@@ -1,0 +1,5 @@
+package com.zelusik.eatery.app.constant.member;
+
+public enum LoginType {
+    KAKAO, APPLE
+}
