+++ src/main/java/com/zelusik/eatery/app/domain/Review.java
@@ -1,7 +1,7 @@
 package com.zelusik.eatery.app.domain;
 
 
-import com.zelusik.eatery.app.domain.constant.ReviewKeyword;
+import com.zelusik.eatery.app.constant.review.ReviewKeyword;
 import com.zelusik.eatery.app.domain.place.Place;
 import com.zelusik.eatery.app.util.ReviewKeywordsConverter;
 import lombok.AccessLevel;
