--- src/main/java/com/zelusik/eatery/app/domain/constant/LoginType.java
@@ -1,5 +1,0 @@
-package com.zelusik.eatery.app.domain.constant;
-
-public enum LoginType {
-    KAKAO, APPLE
-}
