+++ src/main/java/com/zelusik/eatery/app/dto/auth/KakaoOAuthUserInfo.java
@@ -1,7 +1,7 @@
 package com.zelusik.eatery.app.dto.auth;
 
-import com.zelusik.eatery.app.domain.constant.Gender;
-import com.zelusik.eatery.app.domain.constant.LoginType;
+import com.zelusik.eatery.app.constant.member.Gender;
+import com.zelusik.eatery.app.constant.member.LoginType;
 import com.zelusik.eatery.app.dto.member.MemberDto;
 
 import java.time.LocalDateTime;
