+++ src/main/java/com/zelusik/eatery/app/dto/member/MemberDto.java
@@ -1,8 +1,8 @@
 package com.zelusik.eatery.app.dto.member;
 
 import com.zelusik.eatery.app.domain.Member;
-import com.zelusik.eatery.app.domain.constant.Gender;
-import com.zelusik.eatery.app.domain.constant.LoginType;
+import com.zelusik.eatery.app.constant.member.Gender;
+import com.zelusik.eatery.app.constant.member.LoginType;
 import com.zelusik.eatery.app.dto.terms_info.TermsInfoDto;
 
 import java.time.LocalDateTime;
