+++ src/main/java/com/zelusik/eatery/app/dto/place/OpeningHoursDto.java
@@ -1,6 +1,6 @@
 package com.zelusik.eatery.app.dto.place;
 
-import com.zelusik.eatery.app.domain.constant.DayOfWeek;
+import com.zelusik.eatery.app.constant.place.DayOfWeek;
 import com.zelusik.eatery.app.domain.place.OpeningHours;
 
 import java.time.LocalDateTime;
